refactor: Change code style
