+++ src/feature/chapter/mainChapter/MainStory.js
@@ -36,10 +36,11 @@ const MainStory = () => {
             </span>
           </p>
           <p>
-            전형사: &ldquo; {user} 탐정님,{" "}
-            <span>어젯밤 미술관에서 람브람스의 대작이 사라졌습니다. </span>
-            범인 윌리(가명)로 추정되는 사람이 열차를 타고 유럽을 빠져나간다는
-            소문을 입수하였습니다. &quot;
+            <span>
+              전형사: &ldquo; {user} 탐정님, 어젯밤 미술관에서 람브람스의 대작이
+              사라졌습니다. 범인 윌리(가명)로 추정되는 사람이 열차를 타고 유럽을
+              빠져나간다는 소문을 입수하였습니다. &quot;
+            </span>
           </p>
           <p>
             수사에 참여하시려면 <span>참여하기 버튼</span>을 클릭해 주세요.
